backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
<<<<<<< HEAD
psycopg2-binary==2.9.9
python-dotenv==1.0.0
dj-database-url==2.1.0
gunicorn==21.2.0
Pillow==10.1.0
=======
python-decouple==3.8
psycopg2-binary==2.9.7
gunicorn==21.2.0
whitenoise==6.6.0
django-environ==0.11.2
>>>>>>> 9b2160aff06b2f4bae5dc4f518d19142922e4498
